src/requirements.txt
MLB-StatsAPI==1.6.2
slack-sdk==3.21.3
#pybaseball==2.2.7
beautifulsoup4
pytz
openai>=1.0.0